Update miniproject1_3numbers.c
Added a Flower Box and added an if else statement so that the code is "dummy proof".  The if part of the if else statement checks to see if the user put in 3 different numbers and the else part of the if else statement is the code that you already had in there.